Create EEPROM.ino
Currently this program can write and read from the AT28C256, but it cannot do Software Lock and Unlock. So I will be working on that. 

I will also try to modify the program so that people can use this script for different configurations in an easier way.

If you have any suggestions please send them.